requirements.txt
Flask==3.0.3
pandas==2.2.3
python-docx==1.1.2
